refactor: consolidate related territories map and list UI.
Merge map/list panels and hover hook into RelatedTerritoriesSection with a single section hook, and share routing entity type helpers.
