+++ src/hooks/territories/useRelatedTerritoriesSection.ts
@@ -1,0 +1,33 @@
+import { useCallback, useState } from "react";
+import { MapEntityType } from "@/types/rankings";
+import { useRelatedTerritoriesMap } from "./useRelatedTerritoriesMap";
+import { useTerritoryListLayout } from "./useTerritoryListLayout";
+
+export function useRelatedTerritoriesSection(
+  items: string[],
+  entityType: MapEntityType,
+  paginationEnabled: boolean,
+) {
+  const mapState = useRelatedTerritoriesMap({ items, entityType });
+  const [hoveredMapName, setHoveredMapName] = useState<string | null>(null);
+
+  const isHovered = useCallback(
+    (mapName: string) =>
+      hoveredMapName?.toLowerCase() === mapName.toLowerCase(),
+    [hoveredMapName],
+  );
+
+  const layout = useTerritoryListLayout(
+    mapState.territories,
+    paginationEnabled,
+    hoveredMapName,
+  );
+
+  return {
+    ...mapState,
+    ...layout,
+    hoveredMapName,
+    setHoveredMapName,
+    isHovered,
+  };
+}
